+++ build_asm.sh
@@ -21,21 +21,6 @@ VASMFLAGS="$VASMFLAGS -showopt"
 rm -rf "$builddir"
 mkdir -p "$builddir"
 
-buildbin() {
-  export builddir
-  (
-    cd "$1"
-    if [ -e "bin.s" ]; then
-      name="$(basename "$1")"
-      [ -d "$builddir/$name" ] || mkdir -p "$builddir/$name" || exit $?
-      echo "Compile $name BIN program"
-      vasmm68k_mot $VASMFLAGS -Fbin -L "$builddir/$name.bin.lst" -o "$builddir/$name/$name.bin" "$srcdir/$name/bin.s" || exit $?
-      [ -e "$builddir/$name/$name.BIN" ] || exit $?
-      echo
-    fi
-  )
-}
-
 buildasm() {
   if [ -e "$1/boot.s" ]; then
     name="$(basename "$1")"
@@ -56,16 +41,25 @@ buildasm() {
 
   if [ -e "$1/tos.s" ]; then
     name="$(basename "$1")"
-    echo "Compile $name TOS program"
+    echo "Compile $name.TOS"
     [ -d "$builddir/$name" ] || mkdir -p "$builddir/$name" || exit $?
     [ -e "$tosdir" ] || mkdir "$tosdir"
     vasmm68k_mot $VASMFLAGS -monst -Ftos -I"$builddir/$name" -L "$builddir/$name.lst" -o "$tosdir/$name.TOS" "$srcdir/$name/tos.s" || exit $?
     [ -e "$tosdir/$name.TOS" ] || exit $?
     echo
   fi
+
+  if [ -e "$1/prg.s" ]; then
+    name="$(basename "$1")"
+    echo "Compile $name.PRG"
+    [ -d "$builddir/$name" ] || mkdir -p "$builddir/$name" || exit $?
+    [ -e "$tosdir" ] || mkdir "$tosdir"
+    vasmm68k_mot $VASMFLAGS -monst -Ftos -I"$builddir/$name" -L "$builddir/$name.lst" -o "$tosdir/$name.PRG" "$srcdir/$name/prg.s" || exit $?
+    [ -e "$tosdir/$name.PRG" ] || exit $?
+    echo
+  fi
 }
 
-for d in "$srcdir"/*; do buildbin "$d"; done
 for d in "$srcdir"/*; do buildasm "$d"; done
 
 # Clean up